feat: add `findCallExpressionsAndIdentifier` util
