+++ codemods/array.of/index.js
@@ -1,5 +1,5 @@
 import { ts } from '@ast-grep/napi';
-import { findNamedDefaultImport } from '../shared-ast-grep.js';
+import { findCallExpressionsAndIdentifier } from '../shared-ast-grep.js';
 
 const MODULE_NAME = 'array.of';
 
@@ -21,30 +21,12 @@ export default function (options) {
 			const root = ast.root();
 			const edits = [];
 
-			const imports = findNamedDefaultImport(root, MODULE_NAME);
-
-			if (imports.length === 0) {
-				return file.source;
-			}
-
-			let identifierName = null;
-			for (const imp of imports) {
-				const nameMatch = imp.getMatch('NAME');
-				if (nameMatch) {
-					identifierName = nameMatch.text();
-					break;
-				}
-			}
+			const { callExpressions, identifierName, imports } =
+				findCallExpressionsAndIdentifier(root, MODULE_NAME);
 
 			if (!identifierName) {
 				return file.source;
 			}
-
-			const callExpressions = root.findAll({
-				rule: {
-					pattern: `${identifierName}($$$ARGS)`,
-				},
-			});
 
 			for (const call of callExpressions) {
 				const argsMatch = call.getMultipleMatches('ARGS');
